FIX: Fixed nested API route behaviour

+++ app/api/category/[id]/route.ts
@@ -2,8 +2,15 @@ import { NextRequest, NextResponse } from "next/server";
 import { CategoryService } from "@/services/CategoryService";
 import { checkAuth } from "@/utils/checkAuth";
 
-export const GET = checkAuth(async ({ params }: { params: Promise<{ id: number }> }) => {
-    const { id } = await params;
+export const GET = checkAuth(async (req: NextRequest, { params }: { params: { id: string } }) => {
+    const id = Number(params.id);
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { message: "Invalid id" },
+        { status: 400 }
+      );
+    }
+
     try {
         const category = await CategoryService.getCategoryById(id);
 
@@ -24,8 +31,15 @@ export const GET = checkAuth(async ({ params }: { params: Promise<{ id: number }
     }
 });
 
-export const PUT = checkAuth(async (req: NextRequest, { params }: { params: Promise<{ id: number }> }) => {
-    const { id } = await params;
+export const PUT = checkAuth(async (req: NextRequest, { params }: { params: { id: string } }) => {
+    const id = Number(params.id);
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { message: "Invalid id" },
+        { status: 400 }
+      );
+    }
+
     const data = await req.json();
 
     try {
@@ -43,8 +57,15 @@ export const PUT = checkAuth(async (req: NextRequest, { params }: { params: Prom
     }
 });
 
-export const DELETE = checkAuth(async ({ params }: { params: Promise<{ id: number }> }) => {
-    const { id } = await params;
+export const DELETE = checkAuth(async (req: NextRequest, { params }: { params: { id: string } }) => {
+    const id = Number(params.id);
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { message: "Invalid id" },
+        { status: 400 }
+      );
+    }
+
     try {
         await CategoryService.deleteCategory(id);
         return NextResponse.json(
