+++ app/api/location/[id]/route.ts
@@ -2,28 +2,44 @@ import { NextRequest, NextResponse } from "next/server";
 import { LocationService } from "../../../../services/LocationService";
 import { checkAuth } from "@/utils/checkAuth";
 
-export const GET = checkAuth(async ({ params }: { params: Promise<{ id: number }> }) => {
-    const { id } = await params;
+export const GET = checkAuth(async (req: NextRequest, { params }: { params: { id: string } }) => {
+    const id = Number(params.id);
+    if (isNaN(id)) {
+      return NextResponse.json(
+        { message: "Invalid id" },
+        { status: 400 }
+      );
+    }
+
     try {
-        const location = await LocationService.getLocationById(id);
-        if (!location) {
-            return NextResponse.json(
-                { message: "Location not found" },
-                { status: 404 }
-            );
-        }
-        return NextResponse.json(location);
-    } catch (err) {
-        console.error(err);
+      const location = await LocationService.getLocationById(id);
+
+      if (!location) {
         return NextResponse.json(
-            { message: "Internal server error" },
-            { status: 500 }
+          { message: "Location not found" },
+          { status: 404 }
+        );
+      }
+
+      return NextResponse.json(location);
+    } catch (err) {
+      console.error(err);
+      return NextResponse.json(
+        { message: "Internal server error" },
+        { status: 500 }
+      );
+    }
+  }
+);
+
+export const PUT = checkAuth(async (req: NextRequest, { params }: { params: { id: string } }) => {
+    const id = Number(params.id);
+    if (isNaN(id)) {
+        return NextResponse.json(
+            { message: "Invalid id" },
+            { status: 400 }
         );
     }
-});
-
-export const PUT = checkAuth(async (req: NextRequest, { params }: { params: Promise<{ id: number }> }) => {
-    const { id } = await params;
     const data = await req.json();
     try {
         const updatedLocation = await LocationService.updateLocation(id, data);
@@ -40,8 +56,16 @@ export const PUT = checkAuth(async (req: NextRequest, { params }: { params: Prom
     }
 });
 
-export const DELETE = checkAuth(async ({ params }: { params: Promise<{ id: number }> }) => {
-    const { id } = await params;
+export const DELETE = checkAuth(async (req: NextRequest, { params }: { params: { id: string } }) => {
+    const id = Number(params.id);
+
+    if (isNaN(id)) {
+        return NextResponse.json(
+            { message: "Invalid id" },
+            { status: 400 }
+        );
+    }
+
     try {
         await LocationService.deleteLocation(id);
         return NextResponse.json(
